Refactor generated column helpers
